+++ src/hotspot/cpu/x86/jeandleUtils_x86.cpp
@@ -1,0 +1,37 @@
+/*
+ * Copyright (c) 2025, the Jeandle-JDK Authors. All Rights Reserved.
+ * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
+ *
+ * This code is free software; you can redistribute it and/or modify it
+ * under the terms of the GNU General Public License version 2 only, as
+ * published by the Free Software Foundation.
+ *
+ * This code is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+ * FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License
+ * version 2 for more details (a copy is included in the LICENSE file that
+ * accompanied this code).
+ *
+ * You should have received a copy of the GNU General Public License version
+ * 2 along with this work; if not, write to the Free Software Foundation,
+ * Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA.
+ *
+ */
+
+#include "llvm/IR/Jeandle/Attributes.h"
+#include "llvm/IR/Jeandle/GCStrategy.h"
+
+#include "jeandle/jeandleUtils.hpp"
+
+void JeandleFuncSig::setup_description(llvm::Function* func) {
+  func->setCallingConv(llvm::CallingConv::Hotspot_JIT);
+
+  func->setGC(llvm::jeandle::JeandleGC);
+
+  func->addFnAttr(llvm::Attribute::NoCfCheck);
+  func->addFnAttr("patchable-function-entry", "5");
+
+  if (UseCompressedOops) {
+    func->addFnAttr(llvm::Attribute::get(func->getContext(), llvm::jeandle::Attribute::UseCompressedOops));
+  }
+}
